Rework rockbob script to hover script
